xe: jit: remove to_cpp<int> for offsets

+++ src/gpu/intel/jit/ir/send_plan.cpp
@@ -1715,7 +1715,7 @@ private:
 
         // TODO: move unaligned portion of offset to block start x
         auto offset = info_.view().tlayout().offset_in_bytes();
-        if (!is_const(offset) || to_cpp<int>(offset) % base_align)
+        if (!is_const(offset) || to_cpp<int64_t>(offset) % base_align)
             return fail_2d("Unsupported base alignment: ", base_align);
 
         if (!base_mod.is_divisible(base_align) != 0)
@@ -2345,9 +2345,9 @@ private:
         return e;
     }
 
-    static int get_offset(const expr_t &e) {
+    static int64_t get_offset(const expr_t &e) {
         auto *ptr = e.as_ptr<ptr_t>();
-        if (ptr) return to_cpp<int>(ptr->off);
+        if (ptr) return to_cpp<int64_t>(ptr->off);
         ir_assert(e.is<var_t>()) << e;
         return 0;
     }
